Rework combo index
