+++ Cargo.toml
@@ -20,10 +20,6 @@ sparse = ["dep:reqwest"]
 sparse-rustls-tls = ["reqwest?/rustls-tls"]
 # Enables native-tls as the tls implementation for the HTTP client. Not recommended.
 sparse-native-tls = ["reqwest?/native-tls"]
-# Enables a convenience `Index` that can be used to access either a remote git or
-# sparse index, with a convenience constructor for crates.io that takes various
-# configuration factors into account
-combo-index = ["git", "sparse", "dep:toml"]
 
 [dependencies]
 # All paths are assumed to be utf-8 for ease of use and implementation
@@ -44,7 +40,7 @@ smol_str = { version = "0.2.0", features = ["serde"] }
 # Laziness
 thiserror = "1.0"
 # cargo config parsing
-toml = { version = "0.7", optional = true }
+toml = "0.7"
 # Faster hashing
 twox-hash = { version = "1.6", default-features = false }
